codes/api/code/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart